cirrus_vga fix save/restore lfb_addr/lfb_end printfs
A couple of printf calls needed casts to (unsigned long) adding to
avoid compiler compliants, following commits
 97e3066c43477c79b56f4afeb08c330fc2c244b7
 ea324bf41565322b370b27253b1a7eaa3cc360b9
(cherry picked from commit 9722cbb9b813974ee2b2149b683485b077daf078)

Conflicts:

	hw/cirrus_vga.c

+++ hw/vga.c
@@ -1736,7 +1736,7 @@ static void vga_draw_graphic(VGAState *s, int full_update)
                 /* ENODATA just means we have changed mode and will succeed
                  * next time */
                 if (errno != ENODATA)
-                    fprintf(stderr, "track_dirty_vram(%lx, %lx) failed (%d, %d)\n", s->lfb_addr + y, npages, err, errno);
+                    fprintf(stderr, "track_dirty_vram(%lx, %lx) failed (%d, %d)\n", (unsigned long)s->lfb_addr + y, npages, err, errno);
             }
         }
